# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: CircleTrade Agent Stack API
  description: |
    The Circle Agent Stack API exposes CircleTrade AI's export intelligence pipeline as
    individually callable micro-agent endpoints. Each endpoint maps to a distinct intelligence
    module: DDTRS (compliance & HS classification), MeridianFlow (route optimisation),
    and TradeMatch (market & tariff intelligence).

    **Authentication**: All `/api/*` endpoints require an `X-API-Key` header.
    Demo key: `ct-demo-key-2026`
  version: 1.0.0
  contact:
    name: CircleTrade AI
    url: https://circletrade.ai
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: product_description is required
        timestamp:
          type: string
          format: date-time
...
    RouteResult:
      type: object
      properties:
        name: { type: string }
        score: { type: integer }
        transit_days: { type: integer }
        cost_index: { type: number }
        port_efficiency: { type: integer }
        political_risk: { type: string }
        regulatory_alignment: { type: string }
        primary_ports:
          type: array
          items: { type: string }
        recommendation: { type: string }
    RouteResponse:
      type: object
      properties:
        request_id: { type: string }
        query:
          type: object
          properties:
            origin: { type: string }
            destination: { type: string }
            product_category: { type: string, nullable: true }
        routes:
          type: array
          items:
            $ref: '#/components/schemas/RouteResult'
        best_route:
          $ref: '#/components/schemas/RouteResult'
        total_routes_found: { type: integer }
        source: { type: string }
        timestamp: { type: string, format: date-time }
...
    ComplianceChecklistItem:
      type: object
      properties:
        item: { type: string }
        status: { type: string, enum: [REQUIRED, RECOMMENDED] }
        action: { type: string }
    TimelinePhase:
      type: object
      properties:
        phase: { type: string }
        weeks: { type: string }
        actions:
          type: array
          items: { type: string }
    ExportPlanResponse:
      type: object
      properties:
        request_id: { type: string }
        plan_id: { type: string }
        company_name: { type: string }
        product: { type: string }
        origin: { type: string }
        target_market: { type: string }
        category: { type: string }
        hs_code: { type: string }
        executive_summary: { type: string }
        recommended_route:
          type: object
          properties:
            name: { type: string }
            score: { type: integer }
            transit_days: { type: integer }
            cost_index: { type: number }
            primary_ports:
              type: array
              items: { type: string }
        market_analysis:
          type: object
          properties:
            market: { type: string }
            region: { type: string }
            ease_of_trade: { type: integer }
            import_growth_pct: { type: number }
            gdp_growth_pct: { type: number }
            category_fit: { type: string }
            tariff_rate_pct: { type: number, nullable: true }
        compliance_checklist:
          type: array
          items:
            $ref: '#/components/schemas/ComplianceChecklistItem'
        tariff_breakdown:
          type: object
          properties:
            category: { type: string }
            destination_code: { type: string }
            tariff_rate_pct: { type: number, nullable: true }
            tariff_label: { type: string }
            estimated_duty_usd: { type: integer }
        risk_assessment:
          type: object
          properties:
            risk_score: { type: integer }
            risk_level: { type: string, enum: [LOW, MODERATE, HIGH, CRITICAL] }
            license_required: { type: boolean }
            dual_use: { type: boolean }
            key_risks:
              type: array
              items: { type: string }
        timeline:
          type: array
          items:
            $ref: '#/components/schemas/TimelinePhase'
        estimated_costs:
          type: object
          properties:
            budget_input_usd: { type: integer }
            logistics_usd: { type: integer }
            compliance_usd: { type: integer }
            tariffs_usd: { type: integer }
            total_estimated_usd: { type: integer }
            notes: { type: string }
        generated_at: { type: string, format: date-time }
security:
  - ApiKeyAuth: []
paths:
...
  /api/route:
    post:
      summary: Optimise trade corridor / shipping route
      tags: [MeridianFlow]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [origin, destination]
              properties:
                origin:
                  type: string
                  example: UK
                destination:
                  type: string
                  example: UAE
                product_category:
                  type: string
                  example: textiles
      responses:
        '200':
          description: Route optimisation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/export-plan:
    post:
      summary: Generate a structured export plan
      tags: [Pipeline]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [company_name, product, origin, target_market]
              properties:
                company_name:
                  type: string
                  example: "Meridian Exports Ltd"
                product:
                  type: string
                  example: "merino wool scarves"
                origin:
                  type: string
                  example: UK
                target_market:
                  type: string
                  example: UAE
                budget:
                  type: number
                  example: 50000
      responses:
        '200':
          description: Structured export plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportPlanResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Market not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...